requirements.txt
hydra_colorlog==1.0.1
wandb==0.10.25
hydra-core==1.0.6
transformers==4.5.0
pandas==0.25.3
omegaconf==2.0.6
scikit-learn==0.22.1
pytorch-lightning==1.2.1
pytorch-metric-learning==0.9.88
